% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eoo.R
\name{get_eoo}
\alias{get_eoo}
\title{Get extent of occurence (EOO).}
\usage{
get_eoo(species_data, return = "sf")
}
\arguments{
\item{species_data}{Output of get_species_data() function. Supports sp and sf objects.}

\item{return}{Object to return. If return = "sf" (default) will return simple features
object, while return = "sp" will return spatial points data frame object.}
}
\value{
A spatial object (sf or sp class).
}
\description{
Get extent of occurence (EOO).
}
\examples{
sal_atra <- get_species_data("Salamandra atra")
sal_atra_eoo <- get_eoo(sal_atra, return = "sp")
}
