% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esh.R
\name{get_esh}
\alias{get_esh}
\title{Get extent of suitable habitat (ESH).}
\usage{
get_esh(species_data, species_eoo, habitat_variable)
}
\arguments{
\item{species_data}{Output of get_species_data() function. Currently only sp object.}

\item{species_eoo}{Output of get_eoo() function. Currently only sp object.}

\item{habitat_variable}{Raster layer. Should be continous variable with same CRS and overlapping extent.}
}
\value{
Spatial object (currently only sp class - polygon)
}
\description{
Retrieve extent of suitable habitat according to a single environmental variable.
Variable should be continous raster (in future add multiple rasters + categorical) variables.
Function removes the species extent of occurence outside of minimum and maximum values of raster variable.
}
\examples{
sal_atra <- get_species_data("Salamandra atra")
sal_atra_eoo <- get_eoo(sal_atra, return = "sp")
# Load elevation data
srtm_italy <- raster("Y:/Mirza_Cengic/Projects/Other/GSoC/srtm_39_03.tif")

sal_atra_esh <- get_esh(species_data = sal_atra,
                        species_eoo = sal_atra_eoo,
                        habitat_variable = srtm_italy)
}
