% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_data.R
\name{get_species_data}
\alias{get_species_data}
\title{Get species data
Retrieve species data from GBIF repository. You can use species binomial name
to query the database. Query is passed to spocc::occ() function (use ... to pass further arguments to this function).}
\usage{
get_species_data(species_name, return = "sf", return_clean = FALSE,
  country = "", ...)
}
\arguments{
\item{species_name}{Binomial species name in format "Genus species".
Character.}

\item{return}{Object to return. If return = "sf" (default) will return simple features
object, while return = "sp" will return spatial points data frame object.}

\item{return_clean}{Logical (Default is FALSE). Return data with cleaned observation. Data cleaning is
performed with scrubr package (functions coord_impossible(),
coord_incomplete() and coord_unlikely()). See
https://ropensci.org/tutorials/scrubr_tutorial/ for more details.}

\item{country}{Crop results with country boundaries. Results might be
influenced by limit argument from spocc::occ() function. Country data is
retrieved from GADM database. See ?raster::getData() for more details. Use
getData('ISO3') to see the available country options. #' @param ...}
}
\value{
A spatial object (sf or sp class)
}
\description{
Get species data
Retrieve species data from GBIF repository. You can use species binomial name
to query the database. Query is passed to spocc::occ() function (use ... to pass further arguments to this function).
}
\examples{
sal_atra <- get_species_data("Salamandra atra")
plot(sal_atra[1])
sal_atra_ita <- get_species_data("Salamandra atra", return_clean = TRUE, country = "ITA", limit = 10000)
plot(sal_atra_ita[1])
}
